% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glbin-lcd.R
\name{glbin_lcd}
\alias{glbin_lcd}
\title{Block coordinate gradient descent for logistic group lasso}
\usage{
glbin_lcd(X, y, offset, index, eps = 1e-04, lambda, lambda.min = 0.01,
  nlambda = 50, dfmax = ncol(X) + 1, verb = 1, std = TRUE, alpha = 1,
  maxit = 1000, stability_eps = 1e-05)
}
\arguments{
\item{X}{covariate/design matrix. intercept is taken to be the first column!}

\item{y}{response 0/1 vector}

\item{offset}{offset terms}

\item{index}{grouping. set 0 or NA for unpenalised terms. if not given, assumes X[,1] is intercept}

\item{eps}{convergence criterion}

\item{lambda}{vector of lambdas}

\item{lambda.min}{fraction of max lambda to go down to}

\item{nlambda}{number of penalties}

\item{dfmax}{max df, stop if reached}

\item{verb}{verbosity}
}
\description{
Like in Breheny and Huang 2009
}

