% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glbin-lcd-c.R
\name{glbin_lcd_c_std}
\alias{glbin_lcd_c_std}
\title{Block coordinate gradient descent for logistic group lasso (c-version) separate std}
\usage{
glbin_lcd_c_std(X, y, offset, index, eps = 1e-04, lambda,
  lambda.min = 0.001, nlambda = 100, dfmax = ncol(X) + 1, verb = 1,
  add_intercept = TRUE, std = TRUE, alpha = 1, maxit = 1000,
  AIC_stop = 0, gc_force = FALSE)
}
\arguments{
\item{X}{covariate/design matrix. Global intercept should not be here.}

\item{y}{response 0/1 vector}

\item{offset}{offset terms}

\item{index}{grouping. set 0 or NA for unpenalised terms. if not given, assumes X[,1] is intercept}

\item{eps}{convergence criterion}

\item{lambda}{vector of lambdas}

\item{lambda.min}{fraction of max lambda to go down to}

\item{nlambda}{number of penalties}

\item{dfmax}{max df, stop if reached}

\item{verb}{verbosity}

\item{add_intercept}{should the global intercept be added. default: TRUE}

\item{AIC_stop}{default 0. After aic has increased this many steps, halt. 0: go to the end of lambda. If used, should be more than 1.}

\item{gc_force}{FALSE. force garbage collection before going c-side? Might free memory for large X.}
}
\description{
Like in Breheny and Huang 2009
}
