% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-penalty-v2.R
\name{max_penalty}
\alias{max_penalty}
\title{Compute the maximum penalty, possibly sparse}
\usage{
max_penalty(X, y, index, offset, add_intercept = TRUE, center, scale,
  alpha = 1, eps = 0.001, maxiter = 1000, verb = 0, use_glm = FALSE,
  ...)
}
\arguments{
\item{X}{design matrix without intercept}

\item{y}{response 0/1}

\item{index}{grouping}

\item{offset}{offset}

\item{add_intercept}{TRUE}

\item{center}{the column means for standardization}

\item{scale}{sd vector of columns for standardization.}

\item{alpha}{lasso vs ridge weight}

\item{use_glm}{use the glm-function?}

\item{...}{passed on to c-methods}
}
\description{
Compute the maximum penalty, possibly sparse
}
\details{
See glbin for details of the parameters
}
