% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-max-penalty-lcd.R
\name{max_penalty_lcd}
\alias{max_penalty_lcd}
\title{Compute the maximum penalty}
\usage{
max_penalty_lcd(X, y, index, offset, add_intercept = TRUE, alpha = 1)
}
\arguments{
\item{X}{design matrix without intercept}

\item{y}{response 0/1}

\item{index}{grouping}

\item{offset}{offset}

\item{add_intercept}{TRUE}

\item{alpha}{lasso vs ridge weight}
}
\description{
Copied from grpreg
}
\details{
See glbin for details of the parameters
}
