% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs-max-penalty-lcd-old.R
\name{max_penalty_lcd_old}
\alias{max_penalty_lcd_old}
\title{Compute the maximal penalty for lcd}
\usage{
max_penalty_lcd_old(X, y, index, offset, alpha = 1, eps = 1e-04)
}
\arguments{
\item{X}{design matrix}

\item{y}{response 0/1}

\item{index}{grouping}

\item{offset}{offset}
}
\description{
Compute the maximal penalty for lcd
}
\details{
See glbin for details of the parameters
}
